pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentictrading"
version = "0.1.0"
description = "Autonomous strategy factory control plane"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy",
  "pandas",
  "pyarrow",
  "requests",
  "yfinance",
  "alpaca-py>=0.30.0",
  "hmmlearn>=0.3.0",
  "python-dotenv",
  "scikit-learn",
  "optuna",
]

[tool.setuptools]
packages = ["factory", "backtest"]
